Move `CSmoothTime` class to separate compilation unit
