+++ CMakeLists.txt
@@ -2133,6 +2133,8 @@ if(CLIENT)
     serverbrowser_http.h
     serverbrowser_ping_cache.cpp
     serverbrowser_ping_cache.h
+    smooth_time.cpp
+    smooth_time.h
     sound.cpp
     sound.h
     sqlite.cpp
